requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.1
Markdown==3.3.6
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
zipp==3.7.0
